requirements.txt
# ============================
# Core API & Web Framework
# ============================
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.22
cryptography==46.0.4

# ============================
# LLM & Embeddings
# ============================
openai==1.97.0
langchain==0.3.26
tiktoken==0.12.0

# ============================
# OCR (AWS Textract)
# ============================
boto3==1.40.16
botocore==1.40.16


# ============================
# OCR (Azure Document Intell)
# ============================
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.35.0

# ============================
# Vector Store & Numerical Ops
# ============================
faiss-cpu==1.13.2
numpy==2.4.1

# ============================
# Graph Database
# ============================
neo4j==5.28.3

# ============================
# NLP (Entity Extraction)
# ============================
spacy==3.8.11
# Download model separately:
# python -m spacy download en_core_web_sm
# stanza.download('en', package='mimic', processors={'ner': 'i2b2'})

# ============================
# Online Search (Fallback)
# ============================
requests==2.32.5

# ============================
# PDF Handling
# ============================
pypdf==6.6.2

# ============================
# Serialization & Utilities
# ============================
pydantic==2.12.5
python-dotenv==1.1.1
tqdm==4.67.1

# ============================
# Logging & Observability
# ============================
structlog==24.1.0

# ============================
# Chainlit for Chat UI
# ============================
chainlit==2.9.6

# ============================
# LangDetect
# ============================
langdetect==1.0.9